+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".simplecamerax.WalletCameraActivity"
+        <activity android:name=".simplecamerax.CameraActivity"
             android:theme="@style/AppNoActionBarTheme"/>
 
         <meta-data
