+++ app/src/main/res/layout/activity_camera.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/black"
-    tools:context=".simplecamerax.WalletCameraActivity">
+    tools:context=".simplecamerax.CameraActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/leftRectangleViewGuideLine"
